requirements.txt
torch>=2.0.0
Pillow>=10.0.0
transformers>=4.30.0
pillow-heif>=0.13.0

